feat: implement phpcpd and eliminate all code duplication
- Add vendor-bin/phpcpd (sebastian/phpcpd ^6.0) via bamarni/composer-bin-plugin
- Add 'phpcpd' script to composer.json targeting lib/ --exclude lib/Vendor
- Add .github/workflows/phpcpd.yml CI workflow (on: pull_request)
- Reduce duplication from 9 clones (510 lines, 3.45%) to 0 clones

Refactorings to eliminate duplication:
- Extract AbstractInteractiveSetupStateStore: shared base for Signal, Telegram
  and GoWhatsApp InteractiveSetupStateStore (prefix injected via getPrefix())
- Extract AbstractMadelineCommand: shared base for all 5 Telegram CLI commands
  (configure options, session-dir/log helpers, buildMadelineApi)
- Extract AbstractChannelAProvider: shared base for SMS and Telegram AProvider
  (setAppConfig, getSettings, idOverride, getProviderId, isComplete,
  getIdFromProviderFqcn with abstract getDriverNamespacePrefix())
- Extract AbstractCatalogFactory: shared base for SMS and Telegram Factory
  (get() with FQCN+id caching; resolveInstanceCacheKey() is abstract)
