+++ lib/Provider/Channel/Telegram/Provider/Drivers/ClientCli/ResetLogin.php
@@ -10,100 +10,31 @@ declare(strict_types=1);
 
 namespace OCA\TwoFactorGateway\Provider\Channel\Telegram\Provider\Drivers\ClientCli;
 
-use danog\MadelineProto\API;
-use danog\MadelineProto\Logger as MadelineLogger;
-use danog\MadelineProto\Settings;
-use danog\MadelineProto\Settings\AppInfo;
-use danog\MadelineProto\Settings\Logger;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /** @psalm-suppress UndefinedClass */
 #[AsCommand(name: 'telegram:reset-login', description: 'Reset Telegram login state for current session directory')]
-class ResetLogin extends Command {
+class ResetLogin extends AbstractMadelineCommand {
 	#[\Override]
 	protected function configure(): void {
-		$this->addOption(
-			'session-directory',
-			's',
-			InputOption::VALUE_REQUIRED,
-			'Directory to store the session files',
-		);
-		$this->addOption(
-			'log-enabled',
-			null,
-			InputOption::VALUE_REQUIRED,
-			'Whether to persist MadelineProto diagnostics log to file (1/0, true/false)',
-			'0',
-		);
-		$this->addOption(
-			'log-file',
-			null,
-			InputOption::VALUE_REQUIRED,
-			'Absolute path to MadelineProto diagnostics log file',
-			'',
-		);
+		parent::configure();
 	}
 
 	#[\Override]
 	protected function execute(InputInterface $input, OutputInterface $output): int {
-		$sessionDirectory = rtrim((string)$input->getOption('session-directory'), '/\\');
-		if ($sessionDirectory === '') {
-			$output->writeln('<error>Error: Session directory is required.</error>');
+		$sessionDirectory = $this->resolveSessionDirectory($input, $output);
+		if ($sessionDirectory === null) {
 			return Command::FAILURE;
 		}
 
-		$logEnabledRaw = strtolower(trim((string)$input->getOption('log-enabled')));
-		$logEnabled = in_array($logEnabledRaw, ['1', 'true', 'yes', 'on'], true);
-		$configuredLogFile = trim((string)$input->getOption('log-file'));
-		$logFile = $logEnabled
-			? ($configuredLogFile !== '' ? $configuredLogFile : $sessionDirectory . '/MadelineProto.log')
-			: '';
-		if (!is_dir($sessionDirectory)) {
-			@mkdir($sessionDirectory, 0700, true);
-		}
-		if (is_dir($sessionDirectory)) {
-			@chdir($sessionDirectory);
-		}
-		if ($logEnabled) {
-			$logDir = dirname($logFile);
-			if ($logDir !== '' && !is_dir($logDir)) {
-				@mkdir($logDir, 0700, true);
-			}
-		}
-		@ini_set('error_log', '/dev/null');
-
-		$appInfo = new AppInfo();
-		$appInfo->setDeviceModel('Nextcloud-TwoFactor-Gateway');
-		if ($apiId = getenv('TELEGRAM_API_ID')) {
-			$appInfo->setApiId((int)$apiId);
-		}
-		if ($apiHash = getenv('TELEGRAM_API_HASH')) {
-			$appInfo->setApiHash($apiHash);
-		}
+		['logEnabled' => $logEnabled, 'logFile' => $logFile] = $this->resolveLogOptions($input, $sessionDirectory);
+		$this->prepareSessionEnvironment($sessionDirectory, $logEnabled, $logFile);
 
 		try {
-			$loggerSettings = (new Logger())
-				->setLevel(MadelineLogger::LEVEL_NOTICE);
-			if ($logEnabled) {
-				$loggerSettings
-					->setType(MadelineLogger::LOGGER_FILE)
-					->setExtra($logFile);
-			} else {
-				$loggerSettings
-					->setType(MadelineLogger::LOGGER_CALLABLE)
-					->setExtra(static function (): void {
-					});
-			}
-
-			$settings = (new Settings())
-				->setAppInfo($appInfo)
-				->setLogger($loggerSettings);
-
-			$api = new API($sessionDirectory, $settings);
+			$api = $this->buildMadelineApi($sessionDirectory, $logEnabled, $logFile);
 			try {
 				$api->logout();
 			} catch (\Throwable) {
