+++ lib/Provider/Channel/Telegram/Provider/Drivers/ClientCli/SendMessage.php
@@ -11,10 +11,8 @@ declare(strict_types=1);
 namespace OCA\TwoFactorGateway\Provider\Channel\Telegram\Provider\Drivers\ClientCli;
 
 use danog\MadelineProto\API;
-use danog\MadelineProto\Logger as MadelineLogger;
 use danog\MadelineProto\RPCErrorException;
 use danog\MadelineProto\Settings;
-use danog\MadelineProto\Settings\Logger;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -22,16 +20,12 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 #[AsCommand(name: 'telegram:send-message', description: 'Send a message via Telegram Client API')]
-class SendMessage extends Command {
+class SendMessage extends AbstractMadelineCommand {
 
+	#[\Override]
 	protected function configure(): void {
+		parent::configure();
 		$this
-			->addOption(
-				'session-directory',
-				's',
-				InputOption::VALUE_REQUIRED,
-				'Directory to store the session files',
-			)
 			->addOption(
 				'to',
 				't',
@@ -43,63 +37,21 @@ class SendMessage extends Command {
 				'm',
 				InputOption::VALUE_REQUIRED,
 				'Message to send',
-			)
-			->addOption(
-				'log-enabled',
-				null,
-				InputOption::VALUE_REQUIRED,
-				'Whether to persist MadelineProto diagnostics log to file (1/0, true/false)',
-				'0',
-			)
-			->addOption(
-				'log-file',
-				null,
-				InputOption::VALUE_REQUIRED,
-				'Absolute path to MadelineProto diagnostics log file',
-				'',
 			);
 	}
 
+	#[\Override]
 	protected function execute(InputInterface $input, OutputInterface $output): int {
-		$sessionDirectory = rtrim((string)$input->getOption('session-directory'), '/\\');
-		if ($sessionDirectory === '') {
-			$output->writeln('<error>Error: Session directory is required.</error>');
+		$sessionDirectory = $this->resolveSessionDirectory($input, $output);
+		if ($sessionDirectory === null) {
 			return Command::FAILURE;
 		}
 
-		$logEnabledRaw = strtolower(trim((string)$input->getOption('log-enabled')));
-		$logEnabled = in_array($logEnabledRaw, ['1', 'true', 'yes', 'on'], true);
-		$configuredLogFile = trim((string)$input->getOption('log-file'));
-		$logFile = $logEnabled
-			? ($configuredLogFile !== '' ? $configuredLogFile : $sessionDirectory . '/MadelineProto.log')
-			: '';
-		if (!is_dir($sessionDirectory)) {
-			@mkdir($sessionDirectory, 0700, true);
-		}
-		if (is_dir($sessionDirectory)) {
-			@chdir($sessionDirectory);
-		}
-		if ($logEnabled) {
-			$logDir = dirname($logFile);
-			if ($logDir !== '' && !is_dir($logDir)) {
-				@mkdir($logDir, 0700, true);
-			}
-		}
-		@ini_set('error_log', '/dev/null');
+		['logEnabled' => $logEnabled, 'logFile' => $logFile] = $this->resolveLogOptions($input, $sessionDirectory);
+		$this->prepareSessionEnvironment($sessionDirectory, $logEnabled, $logFile);
 
 		try {
-			$loggerSettings = (new Logger())
-				->setLevel(MadelineLogger::LEVEL_NOTICE);
-			if ($logEnabled) {
-				$loggerSettings
-					->setType(MadelineLogger::LOGGER_FILE)
-					->setExtra($logFile);
-			} else {
-				$loggerSettings
-					->setType(MadelineLogger::LOGGER_CALLABLE)
-					->setExtra(static function (): void {
-					});
-			}
+			$loggerSettings = $this->buildMadelineLoggerSettings($logEnabled, $logFile);
 
 			$settings = (new Settings())
 				->setLogger($loggerSettings);
